Move `LocationInfoCache` into `ServiceEndpoint`
