+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/ServiceInstance.kt
@@ -1,27 +1,22 @@
 package net.mullvad.mullvadvpn.service
 
 import android.os.Messenger
+import net.mullvad.mullvadvpn.service.endpoint.LocationInfoCache
 import net.mullvad.mullvadvpn.service.endpoint.SettingsListener
 import net.mullvad.mullvadvpn.util.Intermittent
-import net.mullvad.talpid.ConnectivityListener
 
 class ServiceInstance(
     val messenger: Messenger,
     val daemon: MullvadDaemon,
     val intermittentDaemon: Intermittent<MullvadDaemon>,
     val connectionProxy: ConnectionProxy,
-    val connectivityListener: ConnectivityListener,
     val customDns: CustomDns,
+    val locationInfoCache: LocationInfoCache,
     val settingsListener: SettingsListener,
     val splitTunneling: SplitTunneling
 ) {
     val accountCache = AccountCache(daemon, settingsListener)
     val keyStatusListener = KeyStatusListener(daemon)
-
-    val locationInfoCache =
-        LocationInfoCache(connectivityListener, settingsListener, intermittentDaemon).apply {
-            stateEvents = connectionProxy.onStateChange
-        }
 
     fun onDestroy() {
         accountCache.onDestroy()
