+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/endpoint/LocationInfoCache.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.service
+package net.mullvad.mullvadvpn.service.endpoint
 
 import kotlin.properties.Delegates.observable
 import kotlinx.coroutines.Dispatchers
@@ -14,18 +14,11 @@ import net.mullvad.mullvadvpn.model.Constraint
 import net.mullvad.mullvadvpn.model.GeoIpLocation
 import net.mullvad.mullvadvpn.model.RelaySettings
 import net.mullvad.mullvadvpn.model.TunnelState
-import net.mullvad.mullvadvpn.service.endpoint.SettingsListener
 import net.mullvad.mullvadvpn.util.ExponentialBackoff
-import net.mullvad.mullvadvpn.util.Intermittent
-import net.mullvad.talpid.ConnectivityListener
 import net.mullvad.talpid.tunnel.ActionAfterDisconnect
 import net.mullvad.talpid.util.autoSubscribable
 
-class LocationInfoCache(
-    val connectivityListener: ConnectivityListener,
-    val settingsListener: SettingsListener,
-    val daemon: Intermittent<MullvadDaemon>
-) {
+class LocationInfoCache(private val endpoint: ServiceEndpoint) {
     companion object {
         private enum class RequestFetch {
             ForRealLocation,
@@ -34,6 +27,9 @@ class LocationInfoCache(
     }
 
     private val fetchRequestChannel = runFetcher()
+
+    private val daemon
+        get() = endpoint.intermittentDaemon
 
     private var lastKnownRealLocation: GeoIpLocation? = null
     private var selectedRelayLocation: GeoIpLocation? = null
@@ -73,18 +69,18 @@ class LocationInfoCache(
     }
 
     init {
-        connectivityListener.connectivityNotifier.subscribe(this) { isConnected ->
+        endpoint.connectivityListener.connectivityNotifier.subscribe(this) { isConnected ->
             if (isConnected && state is TunnelState.Disconnected) {
                 fetchRequestChannel.sendBlocking(RequestFetch.ForRealLocation)
             }
         }
 
-        settingsListener.relaySettingsNotifier.subscribe(this, ::updateSelectedLocation)
+        endpoint.settingsListener.relaySettingsNotifier.subscribe(this, ::updateSelectedLocation)
     }
 
     fun onDestroy() {
-        connectivityListener.connectivityNotifier.unsubscribe(this)
-        settingsListener.relaySettingsNotifier.unsubscribe(this)
+        endpoint.connectivityListener.connectivityNotifier.unsubscribe(this)
+        endpoint.settingsListener.relaySettingsNotifier.unsubscribe(this)
         stateEvents = null
 
         fetchRequestChannel.close()
